pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epd-benchmark"
version = "0.1.0"
description = "Benchmark suite for comparing vLLM EPD and ElasticMM disaggregated inference"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
epd-bench = "epd_benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
epd_benchmark = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
